pyproject.toml
[tool.poetry]
name = "validator-schema-excel"
version = "0.1.0"
description = ""
authors = ["Rafael Jürkfitz <rafaeljurkfitz@hotmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.1"
openpyxl = "^3.1.2"
pytest = "^8.1.1"
isort = "^5.13.2"
taskipy = "^1.12.2"
mkdocs = "^1.5.3"
mkdocstrings-python = "^1.10.0"
pygments = "^2.17.2"
pymdown-extensions = "^10.8"
mkdocs-bootstrap386 = "^0.0.2"
pre-commit = "^3.7.0"
pip-audit = "^2.7.2"
pydocstyle = "^6.3.0"
mkdocs-material = "^9.5.18"
black = "^24.4.0"
selenium = "^4.19.0"
streamlit = "^1.33.0"
pandas = "^2.2.2"
numpy = "^1.26.4"
pydantic = {extras = ["email"], version = "^2.7.0"}
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
sentry-sdk = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
known_third_party = []

[tool.taskipy.tasks]
run = "streamlit run app/app.py"
test = "pytest -v"
format = "black . && isort ."
kill = "kill -9 $(lsof -t -i :8000)"
